Fix broken QBluetoothLocalDevice::isValid() behavior
The Bluez and QNX backend always returned true. They should only
return true if the local device is truly available or if the
passed QBluetoothAddress is indeed one of the local bluetooth
devices.

The related unit test was extended to ensure common bahvior of
invalid QBluetoothLocalDevice instance on all platforms.

Task-number: QTBUG-32068
Change-Id: I40ab4db48dc82ba0d1c0bb5275e96a3ee812b01a
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>
Reviewed-by: Alex <alexander.blasche@digia.com>