Extend and improve the unit test for QBluetoothServiceInfo
Note that this change causes  a behavior change. If a service object is
a copy of antoher one they both share the same data and behavior.
Changing one will affect the copy in the same way. No automatic detach
will happen.

THis is a much more expected behavior and avoids issues where one service
info copy was not able to deregister the service anymore.

Change-Id: I0b10688a5ef55e8c1dd9f5ccbf090336e3f635ad
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>