Simplify QBluetoothTransferManager autotest
I don't think the service discovery in there does anything valuable.
Basically every device has a OPP service running.
Like this it is just time consuming and error-prone.

It would also be good to have a tool that automatically accepts the transfer on the testserver so that we can actually execute the test automatically.

Change-Id: I93b3bab32cecb8a3005cecb95a69a40813663e4d
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>