Fix missing QBluetoothSocket::state update after unsuccessful SDP search
When connectToService() triggers an SDP discovery the search may
fail. At the beginning of the discovery the socket state was set to
ServiceLookupState. This patch ensures that the socket state resets to
Unconnected if the SDP discovery failed.

Currently only Bluez uses this discovery mechanism and therefore is the
only affected platform.

Change-Id: I982dafc1f5466071bbf910ed3cf7cf7abda14fe4
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>