Android: Improve QBluetoothLocalDevice::connectedDevices()
There is no API to get the list of connected Bluetooth devices on
Android. However there are service API's which return a list
of their current clients. In this particular case we find the
GATT related service connections. We use that info to complement
the internal list of connected devices.

Note that although this is using Bluetooth Low Energy related API's
the Qt API does not support this feature at this point in time.

Task-number: QTBUG-36764
Change-Id: Ia8266f3a9d24fd16afdd636ca2d3d0a0c6b2b499
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>