Build snippet outside of example build configurations too.
This is a partial revert of 9615e7fbf439d0b2b9476ea7cb15ad145e675583.

The CI does not test a combined developer-build and example
configuration. Ensuring the snippet compiles is mostly a private test.
This patch drops the additional example condition to provide an
increased chance of compiling the snippets on the CI.

So far the condition never triggered and hence the CI couldn't ensure
a compiling snippet.

Task-number: QTBUG-37638
Change-Id: I8d815c6ddeaa2955c2824697a937d3b08729b979
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>