Bluez5: QBluetoothLocalDevice becomes invalid if adapter disappears
Bluez5 supports org.freedesktop.DBus.ObjectManager interface which
permits monitoring of removal/addition of Bt adapters. This permits
the invalidation of the current QBluetoothLocalDevice instance. Once the
instance is invalid, it will remain invalid. Otherwise we could have a
switch of the adapter without the developer noticing it because the next
Bt dongle could get the same interface path. This would
violate the sense of identity that is attached to a
QBluetoothLocalDevice instance.

Task-number: QTBUG-38399
Change-Id: I007278ff1f94c48d6e6902959132088858b0d011
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>