Create 2nd version of QLowEnergyController and add related unit test
The Bluez implementation is the only working implementation for BTLE.
However the current approach of using gatttool via QProcess is not
viable going forward. The tool output is not standardized, the Bluez
developers don't recommend it for anything but debugging and some
Linux distros don't even ship it. Since Bluez does not have any high
level API for BTLE GATT we have to use L2CP sockets to communicate
with the GATT server of the peer device.

The current QLowEnergyController API needs a lot of API changes
going forward as some elements cause problems in the long term.

In summary we need a new internal implementation and a new API for
QLowEnergyController. It is far too much effort to change both elements
at the same time. Hence QLowEnergyControllerNew was added with first
simple API elements. Ultimately it should replace the existing class.
For now the old API stays as is for reference purposes. All other
BTLE classes work just fine for now.

Last but not least a unit test was added. The test assumes the existence
of the TI Sensor tag and verifies all its known characteristics
and descriptors. Ultimately it will help to verify the proper
detection of all details. Right now the test detects all services
and characteristics but not all descriptors.

Change-Id: I6ff1a16da1604be57f5564ae99919ffe7187d02e
Reviewed-by: Lars Knoll <lars.knoll@digia.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>