Remove usage of BluetoothSocket.isConnected()
It seems not all devices reliably return the correct value when calling
isConnected(). connect() succeeds but isConnected() is still false.
On the positive side it is not really needed either.
Calling connect() returns an exception if it fails and close() can be
safely called multiple times. Therefore we only have to check for ocurring
exceptions.

Furthermore isConnected() is API lvl 14 and was not guarded. Removing
the call resolves this problem too.

Task-number: QTBUG-39815
Change-Id: Idf8472c268231a2cfac0f5cd87294384ff779c3f
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>