Share the native private code of QBluetoothLocalDevice
It makes sense to share the private code for QBluetoothLocalDevice
with other classes.

For this purpose the private class QBluetoothLocalDevicePrivate
inherits from the new QBluetoothLocalDevicePrivateData class.

Now QBluetoothLocalDevicePrivateData can be used
by QBluetoothServiceDiscoveryAgentPrivate too.

It reduces code duplication then searching for the handle of
a local device with a certain address.

Change-Id: I0bea137b231553364af9658d7d2a409fef4f40bb
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>