Characteristic and descriptor value() are no longer hex encoded
There is no point converting the data to its hex representation. The
use case at hand will decide it in the current application context.
The returned QByteArray is the raw byte array as it comes from the
device.

This behavior was legacy (introduced by the previous API) and
subsequently it was simply adopted.

Change-Id: If662c02a5e3c3d37cccb374add02e75522352894
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>