Rename descriptorChanged() to descriptorWritten()
This is the part of a change to be able to distinguish
characteristic write confirmations from characteristic
change notifications. Since descriptors
cannot receive notifications they won't have a
changed() signal and all existing
descriptorChanged() signals become descriptorWritten()
signals.

The changed() signal is emitted when a value notification
is received from the device. The written() signal is emitted
when a write request from the API user was successful.

Change-Id: Ief2491ead4723eaa1bca65a09a09c34458f33630
Reviewed-by: Lars Knoll <lars.knoll@digia.com>