Fixing crash when stopping poll loop
When stopping the neard adapter poll loop without
starting it from the same QNearFieldManagerPrivateImpl
instance the m_adapter variable does not hold a
valid pointer to an instance of OrgNeardAdapterInterface.

This patch fixes that issue by creating a temporary
instance on the stack to call the appropriate function
on the D-Bus interface.

Change-Id: I42ced8d00ca9f87756c8f3f81ca78969702344c1
Reviewed-by: Martin Leutelt <martin.leutelt@basyskom.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>