Add QBLuetoothDeviceDiscoveryAgent::UnsupportedPlatformError
iOS does not support classic Bluetooth device discovery. However since
iOS 5.0 it is possible to search for Bluetooth Low Energy devices. On
pre 5.0 devices tLE scan is not possible and this error enum is going
to be used.

In addition the enum is utilized by the default backend
which is used on unsupported Bluetooth platforms.

[ChangeLog][QtBluetooth][QBluetoothDeviceDiscoveryAgent] A new
UnsupportedPlatformError enum value added.

Change-Id: I74b85782ba1e8ac55798171151b364d5dc0aa79f
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>