Workaround Android when connecting to non-existing/turned off device
In such cases a disconnect() call never receives a system callback that
BluetoothGatt is disconnected. Similarly when the timeout hits
the platform sends an unknown error code 133 and again doesn't send
a disconnected status.

This patch works around both issues by preempting the platform behavior.
It seems that subsequent connect attempts are not harmed by this.

Change-Id: I14326c9169f8c1cbbe8aa11de62ca807e79f4975
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>