QLowEnergyController - connectTo/disconnectFromDevice (OS X, iOS)
Fix for disconnectFromDevice.

It can happen that connect requesting is pending forever and not reporting
any error: there is no timeout in Core Bluetooth (quite reasonable),
a device (after it was found by LE scan) can be turned off or move out
of range before connected. If a user calls disconnectFromDevice on a controller
during this time:
- we still have to call cancelPeripheralConnection to make sure the pending
  connection never succeeds
- we have to set the controller's state to unconnected, since, unfortunately,
  delegate's callback is not guaranteed to work (descipe of Apple's docs
  promising this behavior).

Fix for connectToDevice - small typo was breaking connectToDevice's logic.

Change-Id: I4040c79a37ff31ecb5fb2296fa0d39b81d26393d
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>