Android: Add characteristic indication support
At this point in time characteristic indication is very much under-
documented. We have to make some guesses along the way notifications work.

Also, it turns out the code already supported indications but was not
very future proof. If the ClientCharacteristicConfiguration is expanded
by future Bluetooth specifications the notifications could have been
enabled despite the two relevant bits not being set. All that was
required was to write {0xFC FF} and the notifications would have been
enabled. This was due to assumption that only the byte array {0x00 0xFF}
being the disabling data set.

Change-Id: I6c54d557f51977418f91baf658b62e1121785029
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>