QBluetoothLocalDevice - bug on OS X (isValid)
If BT adapter is off local device is invalid on OS X: no way to
access information like address/name, no way to turn the device on,
all operations fail. Considering it's valid breaks the logic:
you can pass any address to such a device (invalid non-null address)
and device can not check if it's valid or not.

Change-Id: I25f70723d566d4de8f993fc9c9a9e6867f936b7f
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>