QBluetoothLocalDevice - isValid, tests (OS X and iOS)
Cancel the previous modification - even if BT is off, the device is
valid unless we create it with some address (in the 'OFF' state I can
not verify any address and the device is considered invalid).
Enable (back) auto-test - if BT is off we just skip the test (as it's done
in other test functions).

Change-Id: I8b013af6a73eab9699a91834e197716836b00e78
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>