Remove the throwing code in Nfc enableForegroundDispatch for Android
In Android Nfc enableForegroundDispatch must not be called while
the activity is not in foreground. We have no API in current Qt
to know if we are in foreground. For that we would need to be
informed when a onPause/onResume in the activity was called.
Then we should enable the dispatch when in foreground and
disable while we are not.

This Patch removes the code that throw if we do it anyway.

Change-Id: I3bff7ba3f23b5ef2665f4c7f164a0bacd3e1c2d2
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>