Bluez5: Fix plain text handover of xml from sdpscanner
The plain text option was never triggered as the raw data processing
always considered \0 as part of the string. Subsequently the string
was always evaluated to be non-printable and the hex based handover
was triggered. Fortunately the hex based handover worked without
trouble.

In addition the plain value was not properly escaped to be valid XML.

Change-Id: I537a94ef4705c7eeab143b3053affc6abbd548cc
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>