Adapt to the new "onResumePause" API
In Anroid the foreground activity get an onResume/onPause
by implemmenting callbacks. This callback will
be called from android when your activity go into
foreground or background. In Nfc we need to know if we
are the activity in foreground for calling enableForegroundDispatch.
In Android this is mandatory.

To support this callback an API was added in qtbase.
This patch implement an listener of this API to get
that events and use some logic to enable and disable foreground
dispatch of nfc tags.

Change-Id: Ibceed51c5d5905a28f01081027fdf5e0e8b4ef12
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>