Android: Fix crash in QBluetoothLocalDevice ctor due to Java exception
In general BluetoothAdapter.getDefaultAdapter should not throw an
exception. If the device does not support Bluetooth the function should
return a null reference only. However some devices throw a
RuntimeException as well. We need to catch it to prevent unwinding of
the stack.

At the same time the patch simplifies the code by using
QAndroidJniObject where possible.

[ChangeLog][QtBluetooth][Android] Fixed crash in QBluetoothLocalDevice
ctor on some devices due to a platform exception in
QBluetoothAdapter.getDefaultAdapter().

Task-number: QTBUG-45066
Change-Id: I933783cda891127e5e37722f96f477c13b433ca7
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>
Reviewed-by: Christian Stromme <christian.stromme@theqtcompany.com>