Fix enum value break
The new values CharacteristicReadError & DescriptorReadError
were introduced in Qt 5.5 whereas all other values of the
QLEService::ServiceError enum were introduced by Qt 5.4.
The new values should not have been put in between the existing enum
values as it breaks binary compatibility.

Although the QLEService API is tech preview in Qt 5.4, there is no
reason to break the compatibility here. In fact this broke the Android
implementation as the enum values are hardcoded on the Java side.

Change-Id: Ic263550f281e03c90f04ae8ae55d1054c1fd7ea6
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>