Android: Avoid code duplication and don't break readWriteQueue
handleFor(Characteristic|Descriptor)() find a handle for a given
char and desc. We should reuse it inside the read notifier for
chars and descs.

In addition, under certain circumstances the read/write queue operation
was finished although the queue was not empty yet. This happened when
the last read operation failed with a certain type of error.
Now we always finish with a call to performNextIO().

Change-Id: Ifdb66c0b03d25ae19db7c49bcbe70eeec2665b6e
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>