Bugfix: Bluetooth Le for Android need to close BluetoothGatt
Android can only handle a defined count of simultaneously connections to bluetooth le
devices. Every Android device have a different defined amount of possible simultaneously connections.
If we do not close the connection and reconnect to the device or another device,
the count of used connections get up. If we exceed the maximum connection count we can not connect
to bluetooth le devices anymore.
The only way to recover is to restart the application to reset the open connection count.

This patch closes the connection after it has been disconnected.

Change-Id: Id96a69ef07eb973216495a0ba8e46fd639338165
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>