Bugfix Andoid Bluetooth coreConfiguration data in discoveredDevices
In current implementation of processDiscoveredDevices, if we receive the same
device again with different coreConfiguration, the device will be just replaced.
In Android there are 2 ways to scan for bluetooth devices. One is the scan for all
bluetooth devices (including LE) and one for scanning just LE devices.
Currently they are used both, which means we receive dual devices multiple times.

Depending of the arrival in time of the QBluetoothDeviceInfo the coreConfiguration is different.
In the List of discovered devices this lead to devices reported as LE only or classic only
which is wrong.

If the Application calls discoveredDevices() it will get a list with wrong coreConfiguration.

This patch update the coreConfiguration of the QBluetoothDeviceInfo in list.

Change-Id: I312544ca513503c5d6b31a1ca0cfbe530c0be911
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>