Bluez: Improve security flag handling in QBluetoothServer
1.) Remove the usage SOL_RFCOMM & SOL_L2CAP. They were replaced
by SOL_BLUEOOTH. The same applies to RFCOMM_LM & L2CAP_LM which were
replaced by BT_SECURITY.

2.) Fix securityFlags() when socket was closed already. It always
triggered undesirable error messages as the code assumed a valid socket
fd.

3.) Fix setSecurityFlags() when calling after the server socket was
closed. The docs state that this function shall be called before
listen is called. However after the socket was closed for the first time
the security level could not be set anymore because the server socket is
invalid until the next call to listen(). To solve this problem the
initial security level is stored and applied during listen().

These changes imply a slightly different runtime behavior of securityFlags().
While the socket is closed it returns the stored value and during
runtime the actual runtime level is returned.

Change-Id: I720e6ac94a569397c4fec0058a1d388a1f35ecae
Reviewed-by: Timur Pocheptsov <Timur.Pocheptsov@digia.com>